+++ solstice/src/main/java/dev/equo/solstice/BuildPluginIdeMain.java
@@ -14,22 +14,74 @@
 package dev.equo.solstice;
 
 import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Enumeration;
+import java.util.List;
 import java.util.function.Function;
+import java.util.stream.Stream;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.BundleException;
 import org.osgi.framework.InvalidSyntaxException;
 
 /**
- * A main method for launching an IDE using Solstice. It supports the following command line
- * arguments
- *
- * <ul>
- *   <li>{@code -installDir C:\SomeDir} determines where the IDE will store its data.
- *   <li>{@code -initOnly true} signals that instead of running an IDE, just activate all bundles
- *       then shut down. Useful for integration testing.
- * </ul>
+ * A main method for launching an IDE using Solstice. It has a verbose command line interface which
+ * is optimized for integration with the EquoIDE Gradle and Maven build plugins, but it can be used
+ * in other contexts as well.
  */
-public class IdeMain {
+public class BuildPluginIdeMain {
+	public enum DebugClasspath {
+		disabled,
+		names,
+		paths;
+
+		public void printWithHead(String header, Stream<String> paths) {
+			switch (this) {
+				case disabled:
+					return;
+				case names:
+				case paths:
+					System.out.println("/ " + header);
+					if (this == names) {
+						paths =
+								paths.map(
+										path -> {
+											int lastSlash = Math.max(path.lastIndexOf('/'), path.lastIndexOf('\\'));
+											return path.substring(lastSlash + 1);
+										});
+					}
+					paths.forEach(System.out::println);
+					System.out.println("\\ " + header);
+					return;
+				default:
+					throw new IllegalArgumentException("Unexpected enum value " + this);
+			}
+		}
+
+		private void printAndExitIfEnabled() throws IOException {
+			switch (this) {
+				case disabled:
+					return;
+				case names:
+				case paths:
+					Enumeration<URL> manifestURLs =
+							SolsticeManifest.class.getClassLoader().getResources(SolsticeManifest.MANIFEST_PATH);
+					List<String> paths = new ArrayList<>();
+					while (manifestURLs.hasMoreElements()) {
+						String url = manifestURLs.nextElement().toExternalForm();
+						paths.add(
+								url.substring(
+										0, url.length() - (SolsticeManifest.SLASH_MANIFEST_PATH.length() + 1)));
+					}
+					printWithHead("jars with manifests inside runtime", paths.stream());
+					System.exit(0);
+				default:
+					throw new IllegalArgumentException("Unexpected enum value " + this);
+			}
+		}
+	}
+
 	private static <T> T parseArg(
 			String[] args, String arg, Function<String, T> parser, T defaultValue) {
 		for (int i = 0; i < args.length - 1; ++i) {
@@ -49,10 +101,14 @@ public class IdeMain {
 		}
 	}
 
-	public static void main(String[] args) throws InvalidSyntaxException, BundleException {
+	public static void main(String[] args)
+			throws InvalidSyntaxException, BundleException, IOException {
 		File installDir = parseArg(args, "-installDir", File::new, defaultDir());
 		boolean useAtomos = parseArg(args, "-useAtomos", Boolean::parseBoolean, false);
 		boolean initOnly = parseArg(args, "-initOnly", Boolean::parseBoolean, false);
+		DebugClasspath debugClasspath =
+				parseArg(args, "-debugClasspath", DebugClasspath::valueOf, DebugClasspath.disabled);
+		debugClasspath.printAndExitIfEnabled();
 
 		BundleContext context;
 		if (useAtomos) {