+++ solstice/src/test/java/dev/equo/solstice/IdeMainBuildPluginTest.java
@@ -13,22 +13,29 @@
  *******************************************************************************/
 package dev.equo.solstice;
 
+import java.io.IOException;
 import org.osgi.framework.BundleException;
 import org.osgi.framework.InvalidSyntaxException;
 import org.slf4j.simple.SimpleLogger;
 
-public class IdeMainTest {
-	public static void main(String[] args) throws InvalidSyntaxException, BundleException {
+public class IdeMainBuildPluginTest {
+	public static void main(String[] args)
+			throws InvalidSyntaxException, BundleException, IOException {
 		boolean useAtomos = true;
 
 		System.setProperty(SimpleLogger.DEFAULT_LOG_LEVEL_KEY, "INFO");
 		System.setProperty(SimpleLogger.LOG_FILE_KEY, "System.out");
 
-		IdeMain.main(
+		BuildPluginIdeMain.main(
 				new String[] {
-					"-installDir", IdeMain.defaultDir().getAbsolutePath(),
-					"-useAtomos", Boolean.toString(useAtomos),
-					"-initOnly", "false"
+					"-installDir",
+					BuildPluginIdeMain.defaultDir().getAbsolutePath(),
+					"-useAtomos",
+					Boolean.toString(useAtomos),
+					"-initOnly",
+					"false",
+					"-debugClasspath",
+					BuildPluginIdeMain.DebugClasspath.names.name()
 				});
 	}
 }