+++ plugin-gradle/CHANGELOG.md
@@ -3,6 +3,9 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+### Added
+- `gradlew equoIde --debug-classpath=[names|paths]` to help debug classpath issues. ([#43](https://github.com/equodev/equo-ide/pull/43))
+  - Also fixed a problem where we were loading multiple versions of some jars as part of our attempted fix in `0.7.1`.
 
 ## [0.7.1] - 2023-01-06
 ### Fixed