+++ plugin-maven/CHANGELOG.md
@@ -3,6 +3,8 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+### Added
+- `equo-ide:launch -DdebugClasspath=[names|paths]` to help debug classpath issues. ([#43](https://github.com/equodev/equo-ide/pull/43)) 
 
 ## [0.7.0] - 2023-01-06
 ### Added